Windows ci stuck (#7)
Generating the ssh key with Perl backticks on Windows lead to a key which did not have an empty passphrase. Subsequently, when trying to use this key, ssh stalls indefinitely waiting on a passphrase to be supplied. The quickest way around this is to generate the test keys with Make, which does manage to invoke `ssh-keygen -N ''` correctly.